Restructure page object
Adapter data was moved into a "docs" key on the page object, and the existing "docs" key was renamed to "body". This keeps documentation data separate from everything else.
The "__fm" key on the page object was renamed to "_frontMatter", and the "_adapterData" key was removed, as the same information can be found in "_frontMatter".